+++ OpenORPG.Server/Game/Movement/MovementSystem.cs
@@ -22,7 +22,7 @@ namespace Server.Game.Movement
         {
         }
 
-        public override void Update(float frameTime)
+        public override void Update(double frameTime)
         {
             
         }
