+++ OpenORPG.Server/Game/World/Systems/DebugMonitorSystem.cs
@@ -20,7 +20,7 @@ namespace Server.Game.World.Systems
             // 
         }
 
-        public override void Update(float frameTime)
+        public override void Update(double frameTime)
         {
             
         }
