+++ OpenORPG.Server/Game/Zones/Spawns/SpawnSet.cs
@@ -15,7 +15,7 @@ namespace Server.Game.Zones.Spawns
     public abstract class SpawnSet
     {
         // The last elapsed spawn time
-        protected float LastSpawnTime = 0f;
+        protected double LastSpawnTime = 0f;
 
         protected int SpawnedSoFar = 0;
